pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.30",
    "numpy>=2.2.6",
    "openai>=1.100.2",
    "opencv-python-headless>=4.12.0.88",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.1",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.1",
    "streamlit>=1.48.1",
    "tesseract>=0.1.3",
]
